+++ imod/tests/test_mf6/test_utilities/test_mf6hfb.py
@@ -3,6 +3,7 @@ import numpy as np
 import pytest
 import shapely
 import xarray as xr
+from pytest_cases import parametrize_with_cases
 
 from imod.mf6.hfb import (
     HorizontalFlowBarrierResistance,
@@ -276,7 +277,70 @@ def test_merge__middle_layer_inactive_domain(
     )
 
     # Assert
-    assert mf6_hfb["cell_id"].shape == (6,)
+    assert mf6_hfb.dataset.coords["cell_id"].shape == (6,)
     assert (mf6_hfb["layer"] == 1).all()
     expected_resistance = single_resistance
     assert (expected_resistance == 1 / mf6_hfb["hydraulic_characteristic"]).all()
+
+
+class InactivityLabelCases:
+    """
+    Labels which will be set to inactive, and expected n_cellids
+    """
+
+    all = slice(None, None)
+
+    def case_all(self):
+        return (self.all, self.all, self.all), 0
+
+    def case_half_left(self):
+        return (self.all, self.all, slice(None, 5.0)), 0
+
+    def case_half_right(self):
+        return (self.all, self.all, slice(5.0, None)), 0
+
+    def case_strip_left(self):
+        return (self.all, self.all, slice(None, 1.0)), 6
+
+    def case_strip_right(self):
+        return (self.all, self.all, slice(9.0, None)), 6
+
+
+@parametrize_with_cases("inactive_labels, n_cellids", cases=InactivityLabelCases)
+@pytest.mark.parametrize("strict_hfb_validation", [True, False])
+@pytest.mark.parametrize("inactive_value", [0, -1])
+def test_merge__single_layer_inactive_domain(
+    modellayers_single_layer,
+    strict_hfb_validation,
+    inactive_value,
+    inactive_labels,
+    n_cellids,
+):
+    """
+    Test with single inactive layer, HFB assigned to that layer should be
+    ignored.
+    """
+    # Arrange
+    single_resistance = 400.0
+
+    modellayers_single_layer["idomain"].loc[inactive_labels] = inactive_value
+
+    hfb_ls = [
+        SingleLayerHorizontalFlowBarrierResistance(
+            make_layer_geometry(single_resistance, 1)
+        ),
+        SingleLayerHorizontalFlowBarrierResistance(
+            make_layer_geometry(single_resistance, 1)
+        ),
+    ]
+
+    # Act
+    if strict_hfb_validation and n_cellids == 0:
+        pytest.xfail("Strict hfb validation and no cellids could be allocated")
+
+    mf6_hfb = merge_hfb_packages(
+        hfb_ls, strict_hfb_validation=strict_hfb_validation, **modellayers_single_layer
+    )
+
+    # Assert
+    assert mf6_hfb.dataset.coords["cell_id"].shape == (n_cellids,)