+++ docs/api/changelog.rst
@@ -33,6 +33,10 @@ Added
 Fixed
 ~~~~~
 
+- Fixed bug where :class:`HorizontalFlowBarrierResistance`,
+  :class:`HorizontalFlowBarrierSingleLayerResistance` and other HFB packages
+  could not be allocated to cell edges when idomain in layer 1 was largely
+  inactive.
 - Fixed bug where :meth:`HorizontalFlowBarrierResistance.clip_box`,
   :meth:`HorizontalFlowBarrierSingleLayerResistance.clip_box` methods only
   returned deepcopy instead of actually clipping the line geometries.